requirements.txt
Django==4.2
django-allow-cidr==0.5.0
django-crispy-forms==1.14.0
django-environ==0.11.2
django-allow-cidr==0.5.0
django-crispy-forms==1.14.0
django-environ==0.11.2
django-mathfilters==1.0.0
django-storages==1.13.1
django-utils-six==2.0
gunicorn==20.1.0
jmespath==1.0.1
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
django-cloudinary-storage==0.3.0
urllib3==1.26.6